backend/requirements.txt
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
boto3==1.39.14
stripe==12.3.0
zappa==0.60.2
python-jose==3.5.0
python-dateutil==2.9.0.post0
cryptography==44.0.0
requests==2.32.4
